Remove need for shell to perform globbing in subprocess.run
The prior method had the shell=True perform the globbing expansion, but
it creates a ruff warning. We perform a first refactor by doing the
globbing ourselves before passing the subprocess.run.
Details for shell expansion: https://stackoverflow.com/a/9997093

We also suppress the Java warning outputs for the PDF conversion.