Move util/ to lib/
The `lib` + `src` structure may be helpful for future refactors.
Specifically, the `util` folder is not a good design pattern as it does
not tell you exactly what it contains.

This folder will also help with extension by perhaps adding `lib/types`
folder that defines what a year, semester, round and student type is.